+++ src/main/java/com/mytiki/ocean/catalog/create/CreateHandler.java
@@ -9,9 +9,9 @@ import com.amazonaws.services.lambda.runtime.Context;
 import com.amazonaws.services.lambda.runtime.RequestHandler;
 import com.amazonaws.services.lambda.runtime.events.APIGatewayV2HTTPEvent;
 import com.amazonaws.services.lambda.runtime.events.APIGatewayV2HTTPResponse;
-import com.mytiki.ocean.catalog.utils.ApiExceptionBuilder;
-import com.mytiki.ocean.catalog.utils.Iceberg;
-import com.mytiki.ocean.catalog.utils.Mapper;
+import com.mytiki.ocean.common.ApiExceptionBuilder;
+import com.mytiki.ocean.common.Iceberg;
+import com.mytiki.ocean.common.Mapper;
 import org.apache.avro.Schema;
 import org.apache.avro.SchemaParseException;
 import org.apache.iceberg.PartitionSpec;
