+++ src/test/java/com/mytiki/ocean/catalog/CreateTest.java
@@ -12,9 +12,9 @@ import com.mytiki.ocean.catalog.create.CreateReq;
 import com.mytiki.ocean.catalog.create.CreateRsp;
 import com.mytiki.ocean.catalog.mock.MockIceberg;
 import com.mytiki.ocean.catalog.mock.MockReq;
-import com.mytiki.ocean.catalog.utils.ApiException;
-import com.mytiki.ocean.catalog.utils.Iceberg;
-import com.mytiki.ocean.catalog.utils.Mapper;
+import com.mytiki.ocean.common.ApiException;
+import com.mytiki.ocean.common.Iceberg;
+import com.mytiki.ocean.common.Mapper;
 import org.junit.Before;
 import org.junit.Ignore;
 import org.junit.Test;
